requirements.txt
# optional
python-twitter==0.8.1
django-taggit==0.9.2
